Jumps -> now 16bit int

+++ SAUCE/famidash.h
@@ -313,7 +313,7 @@ uint8_t kandokidshack2;
 uint8_t kandokidshack3;
 
 uint16_t exittimer;
-uint8_t jumps;
+uint16_t jumps;
 uint8_t orbed[2];
 uint8_t speed;
 uint8_t shuffle_offset;
